+++ CHANGELOG.md
@@ -57,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rule to detect incorrect use of `DISTINCT`. Thanks [@barrywhart](https://github.com/barrywhart).
 - Security fixes from DeepCover. Thanks [@sanketsaurav](https://github.com/sanketsaurav).
 - Automatic fix testing, to help support the newer more complicated rules.
+- Interval literals
 
 ## [0.2.4] - 2019-12-06